add separate functions for parsers
