+++ optionsParsers/basicParser.js
@@ -1,0 +1,8 @@
+const parsArgs = (name) => {
+    const valueIndex = process.argv.indexOf(name);
+    if (~valueIndex) {
+        return process.argv[valueIndex + 1];
+    }
+    return false;
+};
+module.exports = parsArgs;
