+++ optionsParsers/configParser.js
@@ -1,0 +1,5 @@
+const parsArgs = require('./basicParser');
+const configParser = ()=>{
+    return (parsArgs('-c') || parsArgs('--config')) || false;
+};
+module.exports = configParser;
