+++ optionsParsers/inputParsers.js
@@ -1,0 +1,5 @@
+const parsArgs = require('./basicParser');
+const inputParser = ()=>{
+    return parsArgs('-i') || parsArgs('--input') || false;
+};
+module.exports = inputParser;
