+++ optionsParsers/outputParses.js
@@ -1,0 +1,5 @@
+const parsArgs = require('./basicParser');
+const outputParser = ()=>{
+    return parsArgs('-o') || parsArgs('--output') || false;
+};
+module.exports = outputParser;
