+++ output.txt
@@ -14,3 +14,5 @@ TTTbbbtttBBB
 VVVdddvvvDDD
 TTTbbbtttBBB
 TTTbbbtttBBB
+TTTbbbtttBBB
+TTTbbbtttBBB
